requirements.txt
aws-cdk-lib==2.173.1
constructs>=10.0.0,<11.0.0
cdk-aws-lambda-powertools-layer==3.8.0
aws_lambda_powertools==3.3.0
